+++ fabric/src/main/java/com/avp/fabric/data/advancement/AdvancementProvider.java
@@ -16,6 +16,7 @@ public class AdvancementProvider extends FabricAdvancementProvider {
 
     @Override
     public void generateAdvancement(HolderLookup.Provider registryLookup, Consumer<AdvancementHolder> consumer) {
-        AlienAdvancementProvider.generateAdvancement(registryLookup, consumer);
+        AlienAdvancementProvider.generateAdvancements(registryLookup, consumer);
+        HumanAdvancementProvider.generateAdvancements(registryLookup, consumer);
     }
 }